合并熙锦的业务功能

* c4-prodmgmt-xj: (2 commits)
  因冲突更改文件名
  合并熙锦的业务功能 Merge branch c4 into c4-prodmgmt-xj (Conflict resolved online)

Merged-by: 睡不着 <2944119776@qq.com>

CR-link: https://codeup.aliyun.com/zero-one-star/zero-awei/zero-one-08mes/change/63

+++ mes-cpp/mes-c4-prodmgmt/domain/dto/prodmgmt/DeleteWorkReportDTO.h
@@ -16,7 +16,7 @@ class DeleteWorkReportDTO : public oatpp::DTO
 	// 数据库的表： pro_feedback
 	// 
 	// 记录ID
-	API_DTO_FIELD_DEFAULT(Int64, recordId, ZH_WORDS_GETTER("prodmgmt.delete.fields.recordId"));
+	API_DTO_FIELD_DEFAULT(Int32, recordId, ZH_WORDS_GETTER("prodmgmt.delete.fields.recordId"));
 };
 
 #include OATPP_CODEGEN_END(DTO)
